+++ tests/auth_test.py
@@ -27,10 +27,10 @@ def test_login_request(credentials, config_dict):
     """
     Tests that login request produces expected tokens.
     """
-    base_url, realm, client_id = config_dict['base_url'], config_dict['realm'], config_dict['client_id']
+    auth_server_url, realm, client_id = config_dict['auth_server_url'], config_dict['realm'], config_dict['client_id']
     username, password = credentials['username'], credentials['password']
     expected_tokens = prepare_tokens(300, 3600, **credentials)
-    tokens = login_request(base_url, realm, client_id, username, password)
+    tokens = login_request(auth_server_url, realm, client_id, username, password)
 
     assert tokens == expected_tokens
     unstub()
@@ -40,11 +40,11 @@ def test_refresh_request(credentials, config_dict, tokens_dict):
     """
     Tests that refresh request produces expected tokens.
     """
-    base_url, realm, client_id = config_dict['base_url'], config_dict['realm'], config_dict['client_id']
+    auth_server_url, realm, client_id = config_dict['auth_server_url'], config_dict['realm'], config_dict['client_id']
     refresh_token = tokens_dict['refresh_token']
     expect_token_is_valid(refresh_token)
     expected_tokens = prepare_tokens(300, 3600, **credentials, previous_refresh_token = refresh_token)
-    result = refresh_request(base_url, realm, client_id, refresh_token)
+    result = refresh_request(auth_server_url, realm, client_id, refresh_token)
 
     assert result == expected_tokens
     unstub()
@@ -54,9 +54,9 @@ def test_refresh_request_handles_expired_token(config_dict, tokens_dict):
     """
     Tests that refresh request is not made when token is expired.
     """
-    base_url, realm, client_id = config_dict['base_url'], config_dict['realm'], config_dict['client_id']
+    auth_server_url, realm, client_id = config_dict['auth_server_url'], config_dict['realm'], config_dict['client_id']
     refresh_token = tokens_dict['refresh_token']
-    result = refresh_request(base_url, realm, client_id, refresh_token)
+    result = refresh_request(auth_server_url, realm, client_id, refresh_token)
 
     assert result is None
     unstub()
@@ -66,11 +66,11 @@ def test_logout_request(credentials, config_dict):
     """
     Tests that logout request succeeds.
     """
-    base_url, realm, client_id = config_dict['base_url'], config_dict['realm'], config_dict['client_id']
+    auth_server_url, realm, client_id = config_dict['auth_server_url'], config_dict['realm'], config_dict['client_id']
     tokens = prepare_tokens(300, 3600, **credentials)
     refresh_token = tokens['refresh_token']
-    expect_logout(base_url, realm, client_id, refresh_token)
-    result = logout_request(base_url, realm, client_id, refresh_token)
+    expect_logout(auth_server_url, realm, client_id, refresh_token)
+    result = logout_request(auth_server_url, realm, client_id, refresh_token)
 
     assert result is True
     unstub()
@@ -80,12 +80,12 @@ def test_raises_client_authentication_error_if_login_fails(credentials, config_d
     """
     Tests that authentication failure at login raises ClientAuthenticationError.
     """
-    base_url, realm, client_id = config_dict['base_url'], config_dict['realm'], config_dict['client_id']
+    auth_server_url, realm, client_id = config_dict['auth_server_url'], config_dict['realm'], config_dict['client_id']
     username, password = credentials['username'], credentials['password']
     prepare_tokens(300, 3600, status_code=401, **credentials)
 
     with raises(ClientAuthenticationError):
-        login_request(base_url, realm, client_id, username, password)
+        login_request(auth_server_url, realm, client_id, username, password)
     unstub()
 
 
@@ -93,13 +93,13 @@ def test_raises_client_authentication_error_if_refresh_fails(credentials, config
     """
     Tests that authentication failure at refresh raises ClientAuthenticationError
     """
-    base_url, realm, client_id = config_dict['base_url'], config_dict['realm'], config_dict['client_id']
+    auth_server_url, realm, client_id = config_dict['auth_server_url'], config_dict['realm'], config_dict['client_id']
     refresh_token = tokens_dict['refresh_token']
     expect_token_is_valid(refresh_token)
     prepare_tokens(300, 3600, status_code=401, previous_refresh_token = refresh_token, **credentials)
 
     with raises(ClientAuthenticationError):
-        refresh_request(base_url, realm, client_id, refresh_token)
+        refresh_request(auth_server_url, realm, client_id, refresh_token)
     unstub()
 
 
@@ -107,12 +107,12 @@ def test_raises_client_authentication_error_if_logout_fails(config_dict, tokens_
     """
     Tests that authentication failure at logout raises ClientAuthenticationError
     """
-    base_url, realm, client_id = config_dict['base_url'], config_dict['realm'], config_dict['client_id']
+    auth_server_url, realm, client_id = config_dict['auth_server_url'], config_dict['realm'], config_dict['client_id']
     refresh_token = tokens_dict['refresh_token']
-    expect_logout(base_url, realm, client_id, refresh_token, status_code=401)
+    expect_logout(auth_server_url, realm, client_id, refresh_token, status_code=401)
 
     with raises(ClientAuthenticationError):
-        logout_request(base_url, realm, client_id, refresh_token)
+        logout_request(auth_server_url, realm, client_id, refresh_token)
     unstub()